Use u64 for file sizes to handle >4GB on 32-bit
The EROFS format stores sizes as u64, but the code was using usize in
several places. On 32-bit platforms, this would silently truncate file
sizes larger than 4GB.

Change BLOCK_SIZE from usize to u16 (it's a format constant that fits
in 16 bits), which allows using .into() for safe conversions to both
usize and u64. Use u64 for size arithmetic in the writer.

For the reverse direction (u64 to usize for memory allocation), use
fallible conversions with proper error messages so 32-bit systems get
a clear error instead of silent truncation when encountering large
files.

+++ crates/composefs/src/erofs/format.rs
@@ -18,7 +18,7 @@ use zerocopy::{
 /// Number of bits used for block size (12 = 4096 bytes)
 pub const BLOCK_BITS: u8 = 12;
 /// Size of a block in bytes (4096)
-pub const BLOCK_SIZE: usize = 1 << BLOCK_BITS;
+pub const BLOCK_SIZE: u16 = 1 << BLOCK_BITS;
 
 /// Errors that can occur when parsing EROFS format structures
 #[derive(Debug)]
